+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\WebProjects\\monke-scrapr\\src\\index.js":"1","D:\\WebProjects\\monke-scrapr\\src\\App.js":"2","D:\\WebProjects\\monke-scrapr\\src\\reportWebVitals.js":"3","D:\\WebProjects\\monke-scrapr\\src\\components\\Dashboard.jsx":"4","D:\\WebProjects\\monke-scrapr\\src\\components\\Sidebar.jsx":"5"},{"size":500,"mtime":499162500000,"results":"6","hashOfConfig":"7"},{"size":192,"mtime":1612029299673,"results":"8","hashOfConfig":"7"},{"size":362,"mtime":499162500000,"results":"9","hashOfConfig":"7"},{"size":355,"mtime":1612034329410,"results":"10","hashOfConfig":"7"},{"size":2891,"mtime":1612034263534,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"1vf891q",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\WebProjects\\monke-scrapr\\src\\index.js",[],["23","24"],"D:\\WebProjects\\monke-scrapr\\src\\App.js",[],"D:\\WebProjects\\monke-scrapr\\src\\reportWebVitals.js",[],"D:\\WebProjects\\monke-scrapr\\src\\components\\Dashboard.jsx",[],"D:\\WebProjects\\monke-scrapr\\src\\components\\Sidebar.jsx",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\WebProjects\\monke-scrapr\\src\\index.js":"1","D:\\WebProjects\\monke-scrapr\\src\\App.js":"2","D:\\WebProjects\\monke-scrapr\\src\\reportWebVitals.js":"3","D:\\WebProjects\\monke-scrapr\\src\\components\\Dashboard.jsx":"4"},{"size":500,"mtime":499162500000,"results":"5","hashOfConfig":"6"},{"size":192,"mtime":1612029299673,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":499162500000,"results":"8","hashOfConfig":"6"},{"size":4203,"mtime":1612036794518,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1vf891q",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\WebProjects\\monke-scrapr\\src\\index.js",[],["19","20"],"D:\\WebProjects\\monke-scrapr\\src\\App.js",[],"D:\\WebProjects\\monke-scrapr\\src\\reportWebVitals.js",[],"D:\\WebProjects\\monke-scrapr\\src\\components\\Dashboard.jsx",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
