+++ src/components/Dashboard.jsx
@@ -1,20 +1,146 @@
 import React from 'react';
-import Sidebar from './Sidebar';
+import { makeStyles } from '@material-ui/core/styles';
+import Drawer from '@material-ui/core/Drawer';
+import CssBaseline from '@material-ui/core/CssBaseline';
 import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import List from '@material-ui/core/List';
+import Typography from '@material-ui/core/Typography';
+import Divider from '@material-ui/core/Divider';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemIcon from '@material-ui/core/ListItemIcon';
+import ListItemText from '@material-ui/core/ListItemText';
+import PeopleIcon from '@material-ui/icons/People';
+import AddIcon from '@material-ui/icons/Add';
+import BuildIcon from '@material-ui/icons/Build';
+import HomeIcon from '@material-ui/icons/Home';
+import Button from '@material-ui/core/Button';
+import Container from '@material-ui/core/Container';
+import Grid from '@material-ui/core/Grid';
+import Paper from '@material-ui/core/Paper';
+
+const drawerWidth = 240;
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: 'flex',
+  },
+  appBar: {
+    width: `calc(100% - ${drawerWidth}px)`,
+    marginLeft: drawerWidth,
+    background: "#49392C"
+  },
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0,
+  },
+  drawerPaper: {
+    width: drawerWidth,
+    background: "#FFF9C2"
+  },
+  // necessary for content to be below app bar FFF7AE A38266
+  toolbar: theme.mixins.toolbar,
+  
+
+  appBarButtons: {
+      marginLeft: 'auto',
+  },
+  appBarSpacer: theme.mixins.toolbar,
+  content: {
+    flexGrow: 1,
+    height: '100vh',
+    overflow: 'auto',
+  },
+  container: {
+    paddingTop: theme.spacing(4),
+    paddingBottom: theme.spacing(4),
+  },
+  paper: {
+    padding: theme.spacing(10),
+    display: 'flex',
+    overflow: 'auto',
+    flexDirection: 'column',
+
+  },
+  fixedHeight: {
+    height: 240,
+  },
+}));
+
+export default function Dashboard() {
+  const classes = useStyles();
+  
+
+  return (
+    <div className={classes.root}>
+      <CssBaseline />
+      
+      <AppBar position="fixed" className={classes.appBar}>
+        <Toolbar>
+          <Typography variant="h6" noWrap>
+            Monke Scrapr
+          </Typography>
 
 
+        
+          <Button className={classes.appBarButtons}>Login</Button>
+        </Toolbar>
+      </AppBar>
 
-class Dashboard extends React.Component {
+      <Drawer
+        className={classes.drawer}
+        variant="permanent"
+        classes={{
+          paper: classes.drawerPaper,
+        }}
+        anchor="left"
+      >
+        <div className={classes.toolbar} />
+        <Divider />
+        <List>
+          {['Home', 'Scraprs', 'Create Scraprs'].map((text, index) => (
+            <ListItem button key={text}>
+              <ListItemIcon>{index === 0 ? <HomeIcon /> : index === 1 ? <BuildIcon /> : <AddIcon />}</ListItemIcon>
+              <ListItemText primary={text} />
+            </ListItem>
+          ))}
+        </List>
+        <Divider />
+        <List>
+          {['User'].map((text, index) => (
+            <ListItem button key={text}>
+              <ListItemIcon>{index === 0 ? <PeopleIcon /> : <PeopleIcon /> }</ListItemIcon>
+              <ListItemText primary={text} />
+            </ListItem>
+          ))}
+        </List>
+      </Drawer>
 
-
-    render(){
-        return(
-            <div>
-                <AppBar></AppBar>
-                <Sidebar></Sidebar>
-            </div>
-        );
-    }
+      <main className={classes.content}>
+        <div className={classes.appBarSpacer} />
+        <Container maxWidth="lg" className={classes.container}>
+          <Grid container spacing={3}>
+            {/* Chart */}
+            <Grid item xs={3} md={3} lg={4}>
+              <Paper elevation={5} className={classes.paper}>
+                Test
+              </Paper>
+            </Grid>
+            {/* Recent Deposits */}
+            <Grid item xs={12} md={4} lg={4}>
+              <Paper elevation={5} className={classes.paper}>
+                Test
+              </Paper>
+            </Grid>
+            {/* Recent Orders */}
+            <Grid item xs={4}>
+              <Paper elevation={5} className={classes.paper}>
+                Test
+              </Paper>
+            </Grid>
+          </Grid>
+        </Container>
+      </main>
+    </div>
+  );
 }
-
-export default Dashboard;
